# packages/car-management-api/src/api/app/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Car Management App API",
    "version": "1.0.0",
    "description": "API for the Car Management WeChat Mini-program"
  },
  "servers": [
    {
      "url": "/api/v1/app"
    }
  ],
  "paths": {
...
    "/tenants/{tenantId}/car-trims": {
      "get": {
        "summary": "Get all car trims for a category",
        "tags": ["Car Trims"],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of car trims",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarTrim"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/car-trims/{id}": {
      "get": {
        "summary": "Get a car trim by ID",
        "tags": ["Car Trims"],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single car trim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarTrim"
                }
              }
            }
          },
          "404": {
            "description": "Car trim not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CarFeature": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "value": { "type": "string" }
        }
      },
...
      "CarTrim": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tenantId": { "type": "string" },
          "name": { "type": "string" },
          "subtitle": { "type": "string" },
          "image": { "type": "string" },
          "configImageUrl": { "type": "string", "nullable": true },
          "badge": { "type": "string", "nullable": true },
          "originalPrice": { "type": "number" },
          "currentPrice": { "type": "number" },
          "features": { "type": "array", "items": { "$ref": "#/components/schemas/CarFeature" } },
          "categoryId": { "type": "string" },
          "displayOrder": { "type": "integer" },
          "isArchived": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}